Get rid of tsc errors
