+++ sdk/node/lib/hash.js
@@ -1,0 +1,85 @@
+/**
+ * Implement hash primitives.
+ * Currently SHA3 is implemented, but needs to also add SHA2.
+ *
+ * NOTE: This is in pure java script to be compatible with the sjcl.hmac function.
+ */
+var sjcl = require('sjcl');
+
+var hash_sha3_256 = function (hash) {
+
+    if (hash) {
+        this._hash = hash._hash;
+    }
+    else {
+        this.reset();
+    }
+};
+
+hash_sha3_256.hash = function (data) {
+    var hashBits = sjcl.codec.hex.toBits(sha3_256(bitsToBytes(data)));
+    return hashBits;
+};
+
+hash_sha3_256.prototype = {
+
+    blockSize: 1088,
+
+    reset: function () {
+        this._hash = sha3_256.create();
+    },
+
+    update: function (data) {
+        this._hash.update(bitsToBytes(data));
+        return this;
+    },
+
+    finalize: function () {
+        var hash = this._hash.hex();
+        var hashBits = sjcl.codec.hex.toBits(hash);
+        this.reset();
+        return hashBits;
+
+    }
+};
+
+var hash_sha3_384 = function (hash) {
+
+    if (hash) {
+        this._hash = hash._hash;
+    }
+    else {
+        this.reset();
+    }
+};
+
+hash_sha3_384.hash = function (data) {
+    var hashBits = sjcl.codec.hex.toBits(sha3_384(bitsToBytes(data)));
+    return hashBits;
+};
+
+hash_sha3_384.prototype = {
+
+    blockSize: 832,
+
+    reset: function () {
+        this._hash = sha3_384.create();
+    },
+
+    update: function (data) {
+        this._hash.update(bitsToBytes(data));
+        return this;
+    },
+
+    finalize: function () {
+        var hash = this._hash.hex();
+        var hashBits = sjcl.codec.hex.toBits(hash);
+        //debug('finalize hashBits:\n',hashBits)
+        this.reset();
+        return hashBits;
+
+    }
+};
+
+exports.hash_sha3_256 = hash_sha3_256;
+exports.hash_sha3_384 = hash_sha3_384;
